Remove the party emblem from the Manifesto card
There is a styling bug that is causing party emblems to appear
stretched, as they are using the same class as the candidate image
but emblems are not always square. We may need a change in the design
system to add a new class for the party emblem images.
In the meantime, simply removing the emblem from the manifesto card.