Fix bug blocking manifestos being imported
Store the ballot PK's in a list as something strange is
happening later on when trying to access objects from the 'ballots'
queryset after the add_local_party method had been called. Doing so
returns None values rather than an instance.
This resulted in the 'elections' queryset also being empty when
accessing to try to create Manifesto's.
To get around this I am storing the PK's in a list which I then use
to build the Election Queryset. However it is unclear to me why the
unusual behaviour accessing the Ballots queryset is occuring