Add timeouts to ballots and polling station requests
Working theory that sometimes requests to EE are taking too long
which causes the gunicorn workers to die. If this happens to too many
workers it may be killing the instance, leading to site downtime. By
default when using requests there is no timeout, so this adds them
to the calls to EE and WDIV.

The hope is that this still gives 'good' requests time to complete
but kills off 'bad' ones before it kills the gunicorn workers.