Move logger DB migrate to code deploy
Previously logger DB was migrated via lambda, which meant that the
application code was behind the database schema, and could lead to
errors e.g. when a non-null field was added.
This moves the migrate to the final stages of a deployment.
I have also changed the dev/stage environmens to point at the dev RDS
instance. This change was made by amending the variables set in
parameter store.